+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- feat: add file handling and export functionality. (`7848c0f`)
 - refactor: update RowActionsMenu to support grouped menu items. (`98748c5`)
 
 ## 1.5.3 - 2026-06-22